chore: use common utils for time format, stop tooltip if exploreTooltip passed
